✨ added separate home widget
